fix(Workspace) : Clear node/edge object after add node/edge dialog box is closed. (#123)
This commit address the issue that a node label is getting updated if an edge label is updated and vice versa.
This also solves one more issue that a new node is getting created if an edge is updated.
As objects were not getting cleared, the update event updates both the node and edge object.