fix(Workspace) : Disable edit label textbox after creating an edge (#126)
This commit address the issue that edit label textbox was not getting disabled even though the node is out of focus after creating an edge.
First select the node, it will be on focus then draw an edge from selected node to another node.
After creating an edge the node will be deselected however the edit label textbox stays enabled.